Из главного класса вынесен класс итератора
