--- conf.ini
@@ -1,47 +1,0 @@
-[settings]
-;threads - количество параллельных потоков. желательно сделать меньше либо равным количеству реальных
-;(без учёта гипертрединга и тд) ядер процессора. 
-;0 - автоопределение, не используйте его если вы не полный нуп, оно косячит.
-;
-;dimension - размер игрового поля. поле квадратное, в переменной длина стороны
-;(неквадратные поля не поддерживаются и хз будут ли). по умолчанию 400. если тормозит - уменьшайте.
-;
-;iterations - количество итераций до окончания исполнения программы. одна итерация равна dimension*dimension циклам
-;(160000 циклов при стандартном dimension=400). не бойтесть ставить много, программа не теряет управления и её
-;в любой момент можно тупо закрыть, нажав на крестик.
-;примерное время выполнения 20-30 минут на 10000 итераций на стандартных настройках(в зависимости от процессора).
-;0 - будет работать пока не сдохнет :D
-;scale - масштаб, скольким пикселям на экране соответствует бактерия при визуализации
-;если вам кажется, что мелко - увеличивайте
-threads=1
-dimension=150
-iterations=100000
-scale=3
-
-[BacUnit]
-;настройки класса
-;
-;общие правила
-;actlim - потенциал :D это количество доступных ячеек для действий. 
-;relsence - насколько может отличаться по цвету сосед, чтобы оставаться родтвенником. 
-;gainbase - количество энергии, получаемой от способности gain(солнцеедение :D)
-;замечу, что gain, как и другие действия, также расходует энергию
-;light - включает механику освещённости(чем дальше от центра, тем меньше даёт gain)
-actlim=20
-relsence=5
-gainbase=5
-light=true
-
-;начальные статы
-;прокачай свою бактерию! сделай её умной, сильной и прилепи на жгутик спойлер!
-;str - сила, каждая единица увеличивает шанс успешной атаки, но также увеличивает энергопотребление
-;end - выносливость, предельное количество энергии, которое может вместить бактерия. достигнув его,
-;	она делится. если не может - дохнет.
-;mut - вероятность мутации, стандартное значение 50 соответствует 5% шансу.
-str=1
-end=100
-mutstat=100
-mutbeh=100
-;начальный шаблон поведения
-;инструкция прилагается(prog.txt)
-behaviour=gain:
